+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-03-21
+
+### Changed
+
+-   change `rustup-msvc` to `rustup` because default toolchain become `msvc`.
+-   remove `CascadiaCode` font. replace it with `Delugia`.
+
 ## 2023-03-09
 
 ### Added