+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Write about `privacy.sexy` in README.
+- Add wiki for additional documents.
 
 ## 2023-07-03