+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-09-20
+
+### Changed
+
+- Update Sophia Script to v6.5.6.
+- Remove `brotli`, `gitui`, and `goodbye-dpi` from `installAll.ps1`.
+
 ## 2023-08-02
 
 ### Changed